+++ flixopt/io.py
@@ -227,12 +227,12 @@ def save_dataset_to_netcdf(
 
     apply_encoding = False
     if compression != 0:
-        if importlib.util.find_spec('netCDF4') is not None:
+        if importlib.util.find_spec('h5netcdf') is not None:
             apply_encoding = True
         else:
             logger.warning(
-                'Dataset was exported without compression due to missing dependency "netcdf4".'
-                'Install netcdf4 via `pip install netcdf4`.'
+                'Dataset was exported without compression due to missing dependency "h5netcdf".'
+                'Install h5netcdf via `pip install h5netcdf`.'
             )
     ds = ds.copy(deep=True)
     ds.attrs = {'attrs': json.dumps(ds.attrs)}
@@ -255,7 +255,7 @@ def load_dataset_from_netcdf(path: str | pathlib.Path) -> xr.Dataset:
     Returns:
         Dataset: Loaded dataset.
     """
-    ds = xr.load_dataset(str(path))
+    ds = xr.load_dataset(str(path), engine='h5netcdf')
     ds.attrs = json.loads(ds.attrs['attrs'])
     return ds